Fixed access rights to Gnome Shell extension install; Added disabled Gnome Shell animations to work custom configuration; Removed media player Gnome Shell extension from work custom configuration.

+++ functions.sh
@@ -468,6 +468,13 @@ function gnomeshellextension()
         return 1
     fi
 
+    sudo chown -R a+r "/usr/share/gnome-shell/extensions/${ext_uuid}" >/dev/null 2>&1
+    if [[ $? -ne 0 ]]
+    then
+        msgfail '[chown]'
+        return 1
+    fi
+
     sudo rm -f /tmp/extension.zip
 
     msgdone
